+++ data/tournament.go
@@ -146,17 +146,18 @@ func GetTournamentMatches(id int) (map[int][]*models.Match, error) {
 			m.id, m.is_finished, m.current_leg_id, m.winner_id, m.is_walkover, m.created_at, m.updated_at, m.owe_type_id, m.venue_id,
 			mt.id, mt.name, mt.description, mm.id, mm.name, mm.short_name, mm.wins_required, mm.legs_required,
 			v.id, v.name, v.description, m.updated_at as 'last_throw', GROUP_CONCAT(DISTINCT p2l.player_id ORDER BY p2l.order) AS 'players',
-			m.tournament_id, tg.id, GROUP_CONCAT(legs.winner_id ORDER BY legs.id) AS 'legs_won'
+			m.tournament_id, tg.id, GROUP_CONCAT(legs.winner_id ORDER BY legs.id) AS 'legs_won', ot.item
 		FROM matches m
 			JOIN match_type mt ON mt.id = m.match_type_id
 			JOIN match_mode mm ON mm.id = m.match_mode_id
 			LEFT JOIN leg l ON l.id = m.current_leg_id
+			LEFT JOIN owe_type ot ON ot.id = m.owe_type_id
 			LEFT JOIN venue v on v.id = m.venue_id
 			LEFT JOIN player2leg p2l ON p2l.match_id = m.id
 			LEFT JOIN leg legs ON legs.id = p2l.leg_id AND legs.winner_id = p2l.player_id
-			JOIN tournament t ON t.id = m.tournament_id
-			JOIN player2tournament p2t ON p2t.player_id = p2l.player_id AND p2t.tournament_id = t.id
-			JOIN tournament_group tg ON tg.id = p2t.tournament_group_id
+			LEFT JOIN player2tournament p2t ON p2t.tournament_id = m.tournament_id AND p2t.player_id = p2l.player_id
+			LEFT JOIN tournament t ON t.id = p2t.tournament_id
+			LEFT JOIN tournament_group tg ON tg.id = p2t.tournament_group_id
 		WHERE t.id = ?
 		GROUP BY m.id
 		ORDER BY m.id DESC`, id)
@@ -174,10 +175,11 @@ func GetTournamentMatches(id int) (map[int][]*models.Match, error) {
 		venue := new(models.Venue)
 		var players string
 		var legsWon null.String
+		var ot null.String
 		err := rows.Scan(&m.ID, &m.IsFinished, &m.CurrentLegID, &m.WinnerID, &m.IsWalkover, &m.CreatedAt, &m.UpdatedAt, &m.OweTypeID, &m.VenueID,
 			&m.MatchType.ID, &m.MatchType.Name, &m.MatchType.Description,
 			&m.MatchMode.ID, &m.MatchMode.Name, &m.MatchMode.ShortName, &m.MatchMode.WinsRequired, &m.MatchMode.LegsRequired,
-			&venue.ID, &venue.Name, &venue.Description, &m.LastThrow, &players, &m.TournamentID, &groupID, &legsWon)
+			&venue.ID, &venue.Name, &venue.Description, &m.LastThrow, &players, &m.TournamentID, &groupID, &legsWon, &ot)
 		if err != nil {
 			return nil, err
 		}